refactor: extract CommandOptions to types file
