+++ src/command.ts
@@ -11,30 +11,9 @@ import {
     type MiddlewareObj,
 } from "./deps.deno.ts";
 import { InvalidScopeError } from "./errors.ts";
+import { CommandOptions } from "./types.ts";
 
 export type MaybeArray<T> = T | T[];
-
-export type CommandOptions = {
-    /**
-     * The prefix used to identify a command.
-     * Defaults to `/`.
-     */
-    prefix: string;
-    /**
-     * Whether the command should only be matched at the start of the message.
-     * Defaults to `true`.
-     */
-    matchOnlyAtStart: boolean;
-    /**
-     * Whether to ignore or only care about commands ending with the bot's username.
-     * Defaults to `"optional"`.
-     *
-     * - `"ignored"`: only non-targeted commands are matched
-     * - `"optional"`: both targeted and non-targeted commands are matched
-     * - `"ignored"`: only targeted commands are matched
-     */
-    targetedCommands: "ignored" | "optional" | "required";
-};
 
 type BotCommandGroupsScope =
     | BotCommandScopeAllGroupChats
