+++ src/mod.ts
@@ -1,2 +1,3 @@
 export * from "./commands.ts";
 export * from "./context.ts";
+export type { CommandOptions } from "./types.ts";
